requirements.txt
accelerate==0.33.0
beautifulsoup4==4.12.3
boto3==1.34.162
botocore==1.34.162
fastapi==0.115.0
langdetect==1.0.9
numpy==1.26.4
pillow==10.4.0
pydantic==2.8.2
pypdf==4.2.0
pytesseract==0.3.10
python-multipart==0.0.9
qdrant-client==1.11.3
rank-bm25==0.2.2
redis==5.0.8
requests==2.32.3
sentence-transformers==3.0.1
sentencepiece==0.2.0
torch==2.3.1
torchvision==0.18.1
transformers==4.44.2
uvicorn[standard]==0.30.6
streamlit==1.41.1
